requirements.txt
chainlit
google-generativeai
langchain_community
